Adds Section After Code Review & First Steps
Adds the section after code review  which goes into what you do after the code review occurs, such as fix changes if any. Adds section first steps explaining to create an issue for the feature or bug you are working on. Fixes error in Code review section as we can sort of quickly fix it which is why its sort of okay but still not ideal.